pyproject.toml
[project]
name = "biotech-research-ingestion"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=25.1.0",
    "beanie>=2.0.1",
    "boto3>=1.40.46,<1.40.62",
    "aioboto3>=15.5.0,<16",
    "deepagents>=0.4.7",
    "fastapi>=0.135.1",
    "httpx>=0.28.1",
    "langchain>=1.2.10",
    "langchain-anthropic>=1.3.4",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.10",
    "langchain-tavily>=0.2.17",
    "langgraph>=1.0.10",
    "langgraph-checkpoint-postgres>=3.0.4",
    "langsmith>=0.7.14",
    "openai>=2.26.0",
    "psycopg[binary]>=3.3.3",
    "pydantic-settings>=2.13.1",
    "pymongo>=4.16.0",
    "python-socketio>=5.16.1",
    "redis>=7.3.0",
    "tavily-python>=0.7.22",
    "temporalio>=1.23.0",
    "uvicorn[standard]>=0.41.0",
    "fastmcp>=3.1.1",
    "llama-index>=0.14.18",
    "llama-index-llms-openai>=0.7.3",
    "llama-index-tools-mcp>=0.4.8",
    "langmem>=0.0.30",
    "neo4j>=6.1.0",
    "langchain-mcp-adapters>=0.2.2",
]


[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pyright]
# Resolve imports from project root (so "from src.research..." works in IDE)
extraPaths = ["."]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["src"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
